Code Improvements
Cleaned up the app.js file substantially as well as moved all the DOM manipulation out of the javascript file...it is now all handled by KnockoutJS in the HTML.

Added a settings pane with just API Key and an option to reset the conversation.

Added smooth scrolling for longer conversations

Forgot to mention on the initial push but both Markdown and Code Syntax Highlighting is supported.