Claude Vision support added & cleanup
Added support for `Claude Vision` image processing.

If a user has a GPT model selected GPT Vision is used automatically with the `vision::` prefix.

If a user has a Claude model selected it will use Claude Vision automatically with the `vision::` prefix.

Added a new `claude-api-access` file to help clean some of this terrible code up.

tweak to the styling of links show in chat bubbles to be readable.

Various small bugfixes for Claude model implementation.