Fixed dynamic message rendering and auto scroll to bottom bugs.
set scroller height back to 100% and instead made message-list match the messages vh property.

Removed all the random `ScrollToBottom()` calls everywhere and centralized the call in the message-items component and trigger it being called via a single watcher.

Also fixes the scrolling to bottom bug with the dynamic render list.