Refactored and redesigned the chat-input component.
Rediced the width over the input box in desktop mode and centered it in the view.

Removed the loading spinner and replaced it with a slowly pulsating border around the input box while in a loading state.

refactored/updated the styles in general to remove unused or repeated values etc..