Updated the message-items component to use the unique message id properties for the key value.
Removed the showing/hiding of the streaming message text element.

Updated so that the actual message object is being updated when a response is being streamed.

This [Bug] Occasionally the Streaming Text Is Not Visible At All for the First Few Messages #100

This fixes the issue [Bug] Currently Streaming Text Is Not Always in View #110

This also fixes the screen/text flashing after a message stream has finished. Instead saves are not noticable to the user aside from the context window length value updating