Added a push to talk mode for interact mode that can be enabled/disabled in general settings.
There is also a new setting for choosing whether to use whisper or not

Added a dynamic noise floor that is a rolling average of the rms value. This should help allow for more consistent use with various mic sensitivites

Fixed bug causing polling for speech to continue forever even after interact mode was closed.