Updated visions requests context to be saved to the conversation and render the uploaded image on the screen.
Previously vision requests were one off requests that did not get added to the conversation context

Updated the messages list UI design to have large left and right paddings to focus the messages interactions to the center of the screen.

Updated the user messages to be uniquely different but also distinct from the LLM response messages.