Integrating PrimeVue further into the application.
Aligned styles of dropdows across the site to use the general site theme

Switched from slider switches to toggle buttons

replaced the side button in the settings dialog on mobile with a floating menu button that expands the model selector.

Added additional icon that opens context menu to quickly delete or start a new conversation

Added edit icon to the user message bubbles to easy acces all the quick actions.